Moved frame creating and parsing into separate file
